+++ HISTORY.rst
@@ -14,6 +14,9 @@ Bug tracker at https://github.com/giampaolo/psutil/issues
   UnicodeDecodeError exceptions. 'surrogateescape' error handler is now
   used as a workaround for replacing the corrupted data.
 - #741: [OpenBSD] fix compilation on mips64.
+- #737: [Windows] when the bitness of psutil and the target process was
+  different cmdline() and cwd() could return a wrong result or incorrectly
+  report an AccessDenied error.
 
 
 3.4.2 - 2016-01-20